changed the folder structure - The functions to add or remove from cart have stayed in the app.tsx file due to not having proper state management
