+++ src/Cart/Cart.tsx
@@ -2,7 +2,7 @@ import CartItem from "../CartItem/CartItem";
 
 import { Wrapper } from "./Cart.styles";
 
-import { CartItemType } from "../App";
+import { CartItemType } from "../Types";
 
 type Props = {
 	cartItems: CartItemType[];
