+++ src/Types/index.ts
@@ -1,0 +1,9 @@
+export type CartItemType = {
+	id: number;
+	category: string;
+	description: string;
+	price: number;
+	title: string;
+	amount: number;
+	image: string;
+};
